# docs/openapi-production.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRAXIS OS Product API",
    "description": "Authenticated decision-analysis and simulation beta; not certified operational guidance.",
    "version": "0.9.0"
  },
  "paths": {
...
    "/v2/decisions/{identifier}/marketing-support": {
      "post": {
        "summary": "Run Marketing Support",
        "operationId": "run_marketing_support_v2_decisions__identifier__marketing_support_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Channel": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "spend": {
            "type": "number",
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Spend"
          },
          "leads": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Leads"
          },
          "qualified_leads": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Qualified Leads"
          },
          "opportunities": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Opportunities"
          },
          "customers": {
            "type": "integer",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Customers"
          },
          "source": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "spend",
          "leads",
          "qualified_leads",
          "opportunities",
          "customers",
          "source"
        ],
        "title": "Channel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketingRequest": {
        "properties": {
          "base_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Base Version"
          },
          "organization": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Organization"
          },
          "company_size": {
            "type": "string",
            "enum": [
              "solo",
              "micro",
              "small",
              "medium",
              "large",
              "enterprise"
            ],
            "title": "Company Size"
          },
          "industry": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Industry"
          },
          "selling_mode": {
            "type": "string",
            "enum": [
              "b2b",
              "b2c",
              "b2g",
              "mixed"
            ],
            "title": "Selling Mode"
          },
          "product_type": {
            "type": "string",
            "enum": [
              "intelligence_subscription",
              "research_report",
              "data_api",
              "analytics_software",
              "consulting",
              "other"
            ],
            "title": "Product Type"
          },
          "product": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Product"
          },
          "target_customer": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Target Customer"
          },
          "value_proposition": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Value Proposition"
          },
          "differentiator": {
            "type": "string",
            "maxLength": 2000,
            "title": "Differentiator",
            "default": ""
          },
          "problem": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Problem"
          },
          "objective": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Objective"
          },
          "period": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Period"
          },
          "unit": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Unit"
          },
          "cohort_reference": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Cohort Reference"
          },
          "billing_model": {
            "type": "string",
            "enum": [
              "recurring",
              "one_time"
            ],
            "title": "Billing Model"
          },
          "price_per_customer": {
            "type": "number",
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Price Per Customer"
          },
          "delivery_cost_per_customer": {
            "type": "number",
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Delivery Cost Per Customer"
          },
          "onboarding_cost_per_customer": {
            "type": "number",
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Onboarding Cost Per Customer"
          },
          "horizon_months": {
            "type": "integer",
            "maximum": 36.0,
            "minimum": 1.0,
            "title": "Horizon Months"
          },
          "monthly_retention": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Monthly Retention"
          },
          "adverse_conversion_drop": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Adverse Conversion Drop"
          },
          "pilot_budget": {
            "type": "number",
            "maximum": 1000000000000.0,
            "minimum": 0.0,
            "title": "Pilot Budget"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/Channel"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Channels"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Assumptions"
          },
          "requirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Requirements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "base_version",
          "organization",
          "company_size",
          "industry",
          "selling_mode",
          "product_type",
          "product",
          "target_customer",
          "value_proposition",
          "problem",
          "objective",
          "period",
          "unit",
          "cohort_reference",
          "billing_model",
          "price_per_customer",
          "delivery_cost_per_customer",
          "onboarding_cost_per_customer",
          "horizon_months",
          "monthly_retention",
          "adverse_conversion_drop",
          "pilot_budget",
          "channels"
        ],
        "title": "MarketingRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}